Program now generates stability map OK (as well as calculating a single point). GUI will need extensive beta testing to eradicate bugs.
Now need to write save data routine and, on the basis of this, rewrite the load data routine. Will also look at loading a default data set on initialisation. 

Menus need to be made functional.

Should then be ready for first version release - will look into methods of doing this.